backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
tensorflow>=2.16.0
Pillow>=9.0.0
python-multipart>=0.0.5
numpy>=1.21.0
requests>=2.28.0